Update the LangugageProvider in sample and template apps
