--- app/containers/LanguageProvider/constants.js
@@ -1,7 +1,0 @@
-/*
- *
- * LanguageProvider constants
- *
- */
-
-export const CHANGE_LOCALE = 'app/LanguageToggle/CHANGE_LOCALE';
