+++ app/containers/LanguageProvider/selectors.js
@@ -1,5 +1,5 @@
 import { createSelector } from '@reduxjs/toolkit';
-import { initialState } from './reducer';
+import { initialState } from './slice';
 
 /**
  * Direct selector to the language domain
