+++ internals/templates/containers/App/selectors.js
@@ -1,4 +1,8 @@
-import { createSelector } from 'reselect';
+/**
+ * App selectors
+ */
+
+import { createSelector } from '@reduxjs/toolkit';
 
 const selectRouter = state => state.router;
