+++ internals/templates/containers/LanguageProvider/index.js
@@ -9,7 +9,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { useSelector } from 'react-redux';
-import { createSelector } from 'reselect';
+import { createSelector } from '@reduxjs/toolkit';
 import { IntlProvider } from 'react-intl';
 
 import { makeSelectLocale } from './selectors';
