+++ internals/templates/containers/LanguageProvider/selectors.js
@@ -1,5 +1,5 @@
-import { createSelector } from 'reselect';
-import { initialState } from './reducer';
+import { createSelector } from '@reduxjs/toolkit';
+import { initialState } from './slice';
 
 /**
  * Direct selector to the language domain
